FIX: potential segfault on API misuse in TLS 1.3
If the user calls Channel::close() _after_ the channel had encountered
a protocol error the library might segfault. Any protocol error causes the
cipher state object (with the key material) to be destroyed. Calling ::close()
on a Channel in this state caused a method call on said destroyed object and
hence segfault.