Replace BOTAN_FFI_DO/RETURNING with BOTAN_FFI_VISIT
The old macros where hard to understand because they couldn't take
advantage of C++17's generic lambdas. As a side-effect this required
an error-prone distinction into ..._DO and ..._RETURNING, to handle
code blocks that return a value or not.

The new macro accepts a fully qualified lambda instead of a partial
code block. Therefore, the compiler can reason about the lambda's
return value and "do the right thing".

Note: With C++20 we'll be able to get rid of this macro entirely,
using std::source_location.