FIX: response to server's first flight might be plaintext alert
Scenario: After receiving a Client Hello the server responds with its
first flight. The client fails to decode the Server Hello and responds
with an Alert. It cannot protect this Alert, as it didn't understand the
server's key negotiation messages and is forced to send the Alert in the
clear. The server needs to expect that and allow unprotected Alerts,
despite it already having derived key material.