Limit the number of builds we use Boost for
Just installing it can be quite expensive (in one instance took 20m on
a Windows build, 2-3m is more typical) and it adds many lines to the
build.

Boost is especially problematic for the amalgamation builds because
plenty of algorithms in a compiler can be O(n^2) or worse with the
number of lines, and we have a very poor ccache hit rate there already
since changing any single line in the library invalidates the cache
for the entire build.