Improve CPUID support for AVX-512
Now we only flag AVX-512 support if we detect a full set of supported
extensions which are included in Ice Lake, Rocket Lake, and Zen4.

For one this establishes a common baseline for code that uses AVX512.
Secondly, and perhaps more important, the inclusion of VBMI2 ensures
that we will never run AVX512 code on older Intel cores, which have
severe AVX512 clocking penalties.

Add xgetbv checks to ensure ymm and zmm registers are saved by the OS.