Fix a bug in SRP6
The crticial issue is that when computing the client-side key, the
exponent a+u*x was reduced modulo p, which is completely incorrect. At
best, we could reduce it modulo the group order, but this is hard to
do given how SRP groups are structured.

This only caused problems when the hash function used is >= half of
the bitlength of the group, since otherwise the reduction was a
no-op. This is why it caused very occasional failures for the Python
tests; it uses the 1024 bit group with SHA-512.

Outside of using something like SHAKE with SRP6, the most likely case
where this would cause problems is exactly the situation for the
Python tests.

GH #3232