Refactor: Simplify/Clarify implementation of WOTS keypair
Before, the WOTS public and private key had two roles. Namely
to hold the the private/public seed and to be able to derive
specific WOTS keys from those seeds. This didn't quite make sense.

Additionally, there was a lot of boiler plate accessors and
a number of constructors that didn't initialize the objects fully
and were dependent on the call sites to set the correct missing
fields.

Finally, WOTS public and private key classes derived from the
Public_ and Private_Key base classes. Albeit technically correct,
this pulled in override obligations and usage quirks unneccesarily.