FIX: failing test due to undefined behaviour
Its undefined in what order the parameters of a function are
evaluated in C++. Side effects in the evaluation can be harm-
ful: E.g. when using the same RNG for two parameters of the
same function. This made the test output dependent on the
platform it was compiled on. -.-