Remove X509_Object::hash_used_for_signature and Path_Validation_Result::trusted_hashes
Instead the hash that was used is determined only by verifying the
signature, and returned only if the signature was otherwise valid.

PVR::trusted_hashes is difficult to implement and is not really necessary;
it would allow a post-facto check on the hashes that were used, but the whole
point of Path_Validation_Restructions is to restrict that from the start.