Add KEM interface to FFI
Also adds functions to query the output lengths of a PK KEM operation

This also fixes a bug in RSA KEM, which on the encryption side did not
pad the shared key with 0s as is required by the spec. Decryption
already padded with zeros, so the effect would be that sometimes the
encryptor and decryptor would not agree on the shared key.