Fix Python wrappers string conversion
If called with a bytestring everything works as expected.

If called with a string that was all ASCII everything worked as expected.

If called with a string that contained multibyte UTF-8 characters,
then the input would be truncated! because we called it as
`foo(_ctype_bits(x), len(x))` and due to the multibyte encoding,
`len(_ctype_bits(x)) > len(x)`

Also fix mac_update which did not accept strings, unlike the hash
update function.

This is hinted in #2996