Rename the BSI X.509 test data to use shorter names
In #2369 failures in the BSI test suite were reported, this turned out to be
because 7Zip extracted the associated test files with truncated names. This
could be replicated using the command line `7z` util on Linux; 7Zip seems to
truncate paths that are longer than 100 characters.

The paths here were over 90 chars, and with the additional `Botan-2.15.0/` prefix
was sufficient to push us over 100 characters total.