TLS: check that Supported Point Formats Extension contains uncompressed
Both RFC 4492 and the newer RFC 8422 require that the Supported Point
Formats Extension the client sends contains the uncompressed format.

RFC 8422 also explicitly requires servers to check this in case the
Supported Groups extension was sent too.

We go further and always check that clients behave correctly,
and abort the handshake with an illegal_parameter alert if not.

Reported in #2754.