Add a testcase from #3544
The difference is input7.der omits the trailing 7 zero bytes which
were causing the problems described in the issue.

Fix the implementation of peek in ber_dec.cpp which failed to account
for the data already read.

Removes input6.pem from asn1_print test data - this was unused.

Removes some invalid test data from x509_dn.vec - this DN was in fact
completely invalid, and the bug in peek caused us to accept it.