Replace bogus custom curve in TLS example
In #1513 some examples for using a custom curve in TLS was added.

This used an apparently unique 320 bit elliptic curve. However the
curve parameters are certainly not valid, since the group order is far
too small. The Hasse bound specfies that for an elliptic curve over a
finite field p, the order will differ from p by at most about twice
the square root of p. Here the order is far smaller than that.

Replace the example curve with the 256-bit NUMS curve from
https://datatracker.ietf.org/doc/html/draft-black-numscurves-02

Add a check in EC_Group::verify_group that the prime and order
are at least plausibly within the Hasse bound.