Correct an error in RFC 6979 ECDSA
The RFC 6979 mode of ECDSA invoked using the shifted/truncated hash,
without reducing modulo the group order. See Section 2.3.4 of RFC 6979
for the definition.

Unfortunately this situation is missed by the official RFC 6979 tests,
and for other curves we simply generated the test vectors ourselves,
which replicated the error.

DSA always reduced correctly.

This doesn't affect interop or security. However the current tests
trip up an alternate (future) implementation of ECDSA where all of the
variables used in the computation are held in integers which are
strictly bounded modulo the group order, rather than generic BigInts.