Reject elliptic curve private keys that are out of range
This was harmless, in that we would effectively reduce the integer
modulo the order. But there is no reason to allow this.

Update some test vectors which accidentally used a key larger than the
group order.

The affected frp256v1 ECDH tests just needed to reduce the key modulo
the order and everything works as expected.

The brainpool RFC 6979 test vectors change because these tests were
generated by feeding the unreduced key to HMAC_DRBG