In NameConstraints track if the trees contain any unknown names
If so we can immediately fail if the extension is critical.

Also this allows returning early from is_permitted since we don't have
to handle the case of looking for critical+unknown permitted subtrees,
which otherwise would force us to scan (and match against) the entire
list of permitted trees.