Make the new EC_Group constructor more strict
Reduce the allowed range to be between 128 and 521 bits, and it must
be either a multiple of 32 bits long, or if the prime is 521 bits then
it must be exactly P-521.

Require that the field is congruent to 3 modulo 4.

Require that the field and order have the same bit length.