Clean up sqrt logic in pcurves
The sqrt function returned zero to indicate a lack of a square root
but this leads to uncertain situations since zero does itself have a
square root.

Fix point decompression to verify that the x value is a quadratic
root; previously we'd accept invalid x coordinates.

Remove is_square; it was only needed for SSWU and it's faster to just
attempt to square root and then use the returned Choice to decide
which square root to use. This improves the performance of hash to
curve by almost 30%.