FIX: potential use-after-free in tls_proxy
If the TLS session to the client was established (when tls_session_activated()
was called), and the connection to the server was also established successfully
(ec in onConnect() callback was not set); but -- in the mean time -- the this-
pointer was deallocated via std::enable_shared_from_this, we end up in a use-
after free situation.

This sporadically appeared in CI but wasn't reproducible locally, see #4112.