Instead of a mul2_vartime_x_mod_order returning a Scalar, perform equality check
@gmaxwell pointed out in a really great comment on #1479 that you
don't need to actually perform a projective->affine conversion in
ECDSA verification, since instead you can project the r value.

However in the current setup that's not possible since the function is
defining as returning the value and then the comparison happens in the
pubkey code. Instead have the expected value be passed down and all
that comes back is a boolean accept or reject. This allows the
project-r optimization.

This also avoids some back and forth with the various type wrappers,
which is a small win on its own.