Fix timer usage in benchmarks
I noticed on a new-to-me Zen2 laptop that the performance of AES/etc as reported
by the speed cmdlet was quite horrible, especially with small input sizes.
However OpenSSL among others reported good results, as did test programs that
directly called BlockCipher.

It turned out the issue was the failure to amortize calls to the system clock
and CPU cycle counter. This has always been a problem but for whatever reason
calling the cycle counter seems to be much more expensive on this machine than
others of my acquiantance, making the problem bad enough to notice. After this
change, the benchmark results match almost exactly what I see with test programs
that time encrypting buffers in a tight loop.