CI: build/test jitterentropy adapter on some configurations
Currently Ubuntu does not ship with a package for the
jitterentropy-library. Therefore, this pulls the sources from the
maintainer's GitHub and builds/installs it from source.

The library is tiny, so there's no significant overhead in time, but of
course the additional build dependency on another repository is a
bummer.