X25519/X448 throw Invalid_Argument on all-zero result
In normal operation of the algorithms this is essentially impossible
but several protocols require to check for an all-zero return value
of an agreement (for instance TLS;  RFC 8422 Section 5.11.).

RFC 7748 allows to (optionally) fail on this condition. This will simplify
the downstream usage of those key agreements as it does not require any
special treatment anymore.