Add TLS 1.3 groups ML-KEM-512, -768 and -1024
Those groups are described in draft-connolly-tls-mlkem-key-agreement and
request registration of code points 0x0200, 0x0201 and 0x0202 from IANA.

Previously, 0x0200 and 0x0202 were experimentally assigned to eFrodoKEM
variants (in compatibility with liboqs). Therefore we had to re-assign
eFRODOKEM_640_AES_OQS and eFRODOKEM_976_AES_OQS to the new code points
used by the latest liboqs version.

This commit implements the groups as described in the ...-05 draft version.