FIX: secret-dependent branch in BigInt::ct_cond_assign()
Both clang and GCC generate a conditional jump on a secret-dependent
value when compiling without any optimizations (-O0)!
My best guess is that the conjunction of two bools is generated
canonically *jumping* over the evaluation of the second opperand when
the first was already found to be false. Optimizers will figure out
that this can be implemented using an 'and' instruction for two bool
operands.