Cleanups relating to CurveGFp
This removes CurveGFp_Repr and its subclasses. Now the Montgomery
logic previously implemented there is delegated to the already
existing Montgomery_Params.

This also guts CurveGFp; at this point it could easily be removed but
unfortunately this type is used in a couple of interfaces which, while
already deprecated, are still part of the public API. So complete
removal will have to wait for Botan4.