During batch affine conversion avoid leaking which point was the identity
This doesn't affect performance in the usual case of no identity elements in the
input, since we then have to examine every point. The lack of early exit might
even improve performance (eg through vectorization), though Clang doesn't seem
to be taking advantage of this at the moment.

I can't think of any situation where this leakage could affect anything; the
only "usual" case it occurs is when performing verification using an ECDSA key
which is chosen as a very small multiple of the group generator. In that case
the inputs are anyway public, and the secret key is also (effectively) public.
But since it's easy and effectively costless to avoid, might as well.