SP.800-108 KDFs can handle different counter and L byte widths
This allows the application to define the encoding bit lengths of the
KDF's counter and output length fields. Bit values 8, 16, 24 and 32 are
supported. Technically, SP.800-108 supports any values between 1 and 32,
but we opted not to support non-byte-aligned values.

By default both fields are defined as 32 bit wide to retain the legacy
behavior if applications don't explicitly specify those values.

Co-Authored-By: Amos Treiber <amos.treiber@rohde-schwarz.com>