Add volatile to word8_add2 inline asm block
Apparently in some situations GCC does not realize that calling this block twice
might have visible results. In the BEEA added in pcurves in #4620, in the
reduction loop we had logic like `if(borrow) x += INV_2` in a loop, and GCC
decided that adding `INV_2` to `x` just the once was sufficient.

Adding volatile was done for the other similar asm statements in 9b1b666f81
after similar issues arose during early development of the pcurves
implementation, but was apparently missed for word8_add2.