Remove pcurves "user interface" API
In the very initial stages of pcurves the idea was that pcurves would
be the new internal interface used for EC arithmetic. For example that
ECDSA would be implemented directly using pcurves. So some effort went
towards providing a "nice"ish API directly from pcurves.h

At least in the short term we can't possible do this (due to needing
to still handle application curves with cofactors or huge parameters),
and in the long term, even as we drop support for such things, it
seems better to take advantage to simplify EC_Scalar and company,
which would have about the same effect on overhead.