Remove an old test related to decoding keys using explicit parameters
This uses a funky curve, apparently "primeCurve 33" from TU-Darmstadt [*], about
which I can find no further information. It uses a 460 bit field, which we will
not support anyway once the BigInt EC arithmetic is removed.

The test is anyway redundant with others so I don't think we need to keep this
around.

https://www.hrz.tu-darmstadt.de/services/it_services/oids/oids_der_informatik_cdc/index.en.jsp