Don't bother with loop blocking for carry chains with constant lengths
As long as the iteration bound is a compile time constant, and not too large [*],
both GCC and Clang will completely unroll a loop calling word_add or word_sub.

Reformulate word_add so that it can use dword if that is helpful. Unfortunately,
GCC generates better code with dword but Clang only recognizes our current
pattern for a carry chain, so we choose at compile time based on `__clang__`.

Improvements vary by compiler and curve. With Clang, improvement ranged from 3% to
17%, with the best improvement being for secp521r1 variable point multiplications.
GCC improvement was more muted, typically 3% to 7% with one standout improvement for
secp521r1 variable point multiplication.

[*] Measured 17 for GCC and 32 for Clang, but this probably varies based on
a variety of factors.