Misc FFI cleanups
Add missing null pointer checks; while we don't guarantee that FFI will never
crash if passed a null pointer we can do our best to notice if it happens.

A number of key loading/generation functions would leave the key parameter
uninitialized if something failed. The typical pattern is to set it to nullptr
on entry and then if successful set it to the new object.

Support capitalized X25519 and X448 in botan_privkey_create_ecdh; it's odd
that we generally support only that naming but required lowercase here.

Clean up the Kyber R3 loading functions, they had a bit of duplication.

Use span where helpful.