Remove support for calling NetBSD _dlauxinfo
This was buggy and would crash due to a mixup of comparing the pointer with
AT_NULL, when it should have been checking the type field. So the search would
walk off the end of the struct if asked for a field that didn't exist at runtime.

This is an easy fix, but it appears that NetBSD's ELF auxvec doesn't provide the
information we are looking for anyway. That being the case, just remove it entirely.

GH #4733