Inline thunk calls in FFI
For compilation using libstdc++ this cuts between 200 and 250 KB off the size of
the shared object. This tracks across both GCC and Clang, using -O3 or -Os, and
on multiple architectures. This is a 2-3% overall size reduction.

For compilation using libc++, this change saves between 800 and 900 KB of code
space in the shared object. Again this seems to track closely across multiple
different architectures and operating systems. This is a 8-9% (!) overall size
reduction.